refactor: move helper functions to helper module
